refactor: move concrete store implementation into sub-package of hwes
