+++ services/tasks-svc/main.go
@@ -4,7 +4,7 @@ import (
 	"common"
 	pb "gen/proto/services/tasks_svc/v1"
 	daprd "github.com/dapr/go-sdk/service/grpc"
-	"hwes"
+	"hwes/eventstoredb"
 	"tasks-svc/internal/task/api"
 	"tasks-svc/internal/task/service"
 )
@@ -17,9 +17,10 @@ var Version string
 func main() {
 	common.Setup(ServiceName, Version, true)
 
-	hwes.SetupEventStoreByEnv()
+	eventstoredb.SetupEventStoreByEnv()
 
-	aggregateStore := hwes.NewAggregateStore(hwes.GetES())
+	eventStore := eventstoredb.SetupEventStoreByEnv()
+	aggregateStore := eventstoredb.NewAggregateStore(eventStore)
 	taskService := service.NewTaskService(aggregateStore)
 
 	common.StartNewGRPCServer(common.ResolveAddrFromEnv(), func(server *daprd.Server) {
